Additional values for MOODLE_DOCKER_PHP_VERSION
I tried setting MOODLE_DOCKER_PHP_VERSION to 7.2 and it did not change the PHP version. Turns out I needed to set it to 7.2-stretch. So I added in additional values that MOODLE_DOCKER_PHP_VERSION can be used to include 7.2-stretch and 7.3-stretch. 

It is not obvious that this is the proper values to change PHP to a version higher than 7.1.